data: biatorbagy: port to simplify=false
Note that 'Május 1. utca' is removed for now: it did not filter anything
(since we have default filters) and "." in a yaml key name triggers
<https://github.com/23andMe/Yamale/pull/51>.

Let's hope that gets merged and released before we need this
functionality. (Failing that, we should just drop yamale and came up
with an own validator.)