Switch from yamale to the own validator
The main benefit is that yamale is reasonably good at checking what *is*
in the data, but bad at checking what is not in the data.

It gives error messages like:

	failed to validate data/relation-sashegy.yaml: unexpected key 'filters.Bod Péter lejtő.ranges[0].start'

i.e. no line numbers; that is a trade-off: either we have line numbers
or we use the same yaml parser as the one used at runtime, I went for
the later.