normalize, non-interpolation=all case: handle parity mismatch of ranges
Fixes <https://github.com/vmiklos/osm-gimmisn/issues/211>.

If this is not an interpolation=all street and we have a range, then
don't expand it if the parity of the start and stop doesn't match.

That's not really a valid range, just a real life oddity, so don't
expand it or interpret it any way, other then the start + stop numbers,
as before.

Tested-by: Travis