Rename helpers.py to areas.py
Given it really just contains the Relation class and associated
functionality. The new module is called "areas" rather than "relations",
as "relations" is an already frequently used local variable name and
such a clash is good to be avoided.

This has the benefit that we again have a single "shared" module in the
form of util.py.

Tested-by: Travis