Street housenumbers reference: use tab as separator
Both the street name and the house number may contain spaces so using
spaces as a separator between them is not a great idea. This was
especially problematic in case one street was "foo" and the other was
"foo bar", because then "foo bar 1" was interpreted as a "bar 1" house
number of "foo".

Fixes <https://github.com/vmiklos/osm-gimmisn/issues/312>.

Tested-by: Travis