percent format: depend on UI language, not locale
This has the benefit that the percent format is consistent with the
translations. The downside is that the locale module only works with the
current per-process locale, so it can't work with the per-thread UI
language.

Depending on ICU just for this is an overkill, so introduce a simple
dictionary for now.

Related to <https://github.com/vmiklos/osm-gimmisn/issues/564>.

Tested-by: Travis