Update list of active relations based on 2020-07 data
Enable the top visited relations (20%), disable the rest: 32 removals
and 19 additions.

The idea is that many relations get added, there is a short-term active
survey and then it wastes CPU on a daily basis forever. So auto-disable
those, and also auto-enable other relations which are popular but
people shy to activate them.

Monthly automatic update + manual updates are still there for all
relations.

Tested-by: Travis