typescript: migrate osm.js
Go with typescript-4.0.3, because my code completion uses that and 4.1
would rename Position to GeolocationPosition and then code editing and
building would be out of sync.

See <https://github.com/DefinitelyTyped/DefinitelyTyped/pull/47681> and
<https://github.com/ycm-core/ycmd/blob/master/third_party/tsserver/package.json>.

Benefit is the same as type hints for Python: I'll be able to understand
this code even if I didn't write it yesterday.

Change-Id: I4d619ce3e1b071d9b76879a450449b31437d5bd6