invalid refstreets: add to missing/additional streets as well
We have a list of invalid osm and ref names for the osm <-> ref street
name mapping. This was already visible at the bottom of the missing
housenumbers page. Do the same for the missing/additional streets pages,
as this is useful for relations where only streets are analyzed.

Addresses <https://github.com/vmiklos/osm-gimmisn/issues/1008>.

Change-Id: I657f7307fff21540f50e6e8be39cd4dd9bbe6435