Cache missing-housenumbers HTML output
Old cost for budapest_11: 26.888 seconds

New cost: 0.054 seconds

I.e. once you hit the cached cache, the new cost is 0.2% of the baseline.

Note that the caches is not versioned, the path can be always extended
to contain a version number later if it's needed.

Also note that the cached content is localized, so the language is part
of the cache key.

Fixes <https://github.com/vmiklos/osm-gimmisn/issues/1119>.

Change-Id: Ia3d11833b247879b6452d47b93c43d7576902269