Cache missing-housenumbers plain text output
Old cost for budapest_11: 28.123 seconds

New cost: 0.053 seconds

I.e. once you hit the cached cache, the new cost is 0.19% of the baseline.

Note that the caches is not versioned, the path can be always extended
to contain a version number later if it's needed.

Also note that the cached content is not localized, so the language is
not part of the cache key.

Addresses <https://github.com/vmiklos/osm-gimmisn/issues/1206>.

Change-Id: Iad51dafcddedcd5dca13d5b59c3f645e59e3b145