cache_yamls: switch from pickle to json
Pickle is a python-specific serialization that allows quickly loading
nested dictionaries from disk.

We use yaml for the in-git filter data as it's a better markup for
manual editing: comments, etc.

Given that yaml provides nice manual editing, it's naturally slower to
load than pickle. Switch to json as the cache format: json is still
hard to edit manually (e.g. no comments possible) but it's not
python-specific and almost as fast as pickle.

(Sure, we could use some non-standard json parser that supports
comments, but that'll be probably slower than the stupid+fast built-in
one.)

Change-Id: I8c1dfc33e6cd6fc0991aaf8795b7202b72cb0fbc