config.FileSystem: get_relation_allowlist -> open
The first is a very specific hook, while the later is a generic one that
we can reuse in other cases.

This moves towards in-memory files during tests. At least not writing
the real filesystem during tests will be needed if we'll want to run the
tests in parallel in the future.

Change-Id: If70d4c23ca624cdf37cab1a78c156ecb937fead5