areas: speed up get_missing_housenumbers()
Now that the relevant piece of this is free from Python, callgrind shows
where is the not needed copying.

Old cost for 'time ./missing_housenumbers.py budapest_11':

real    0m12,599s

New cost:

real    0m0,703s

The best Python time was 0m0,943s before we started to pay for the
Python<->Rust copying, so finally this is some real benefit.

Change-Id: I6095714e64860ec2daf58ce73ccc92a5bdb8d365