Move the TypeScript code to src/browser/
Initially everything was in the root dir, but Rust code is now under
src/, so move the TypeScript code there as well. This way one can easily
grap for TODOs in data/ and src/ separately.

Also fix a bug where the stats page wanted to show a chart for the ZIP
progress stat.

Change-Id: Icaab0e5ae92dffa417b691ec866b188e0aeab9c3