get_sort_key: simplify error handling
The root of the problem is that we compare unicode strings with icu,
where API calls can fail, but rust wants a comparator that never fails.

This was solved by adding unwrap() around all uses of get_sort_key(),
but this is ugly.

Instead, it turns out that ucol::UCollator::try_from() ~never fails in
practice (even if you pass garbage to it) and we know that the input
string is valid utf8 (comes from Rust), so ustring::UChar::try_from()
won't fail in our case, either.

This means we can just use expect() around those two calls and then
get_sort_key() never fails, matching the expectations from Rust.

Change-Id: I157de8619e4c2686ac9023dc0f08ad81818314e7