Introduce util::get_tz_offset()
We want to display time in the local timezone, but localtime_r() is not
safe to be called on a thread.

So get the local tz offset using a lazy static and call it before
rouille starts, which is the only case where we're multithreaded.

This also allows returning a UTC timestamp in context::system and the
callers can apply to_offset(util::get_tz_offset()) on the parsed result
if they want to display the date in local time.

Change-Id: I199dc5d9b11e4ec928c88dfbd47b023c829b1978