context: lazy-create a shared database connection
Right now all 3 places create their own sql connections, which is not a
big problem, but in case more sql usage is introduced in the future,
then this connect / disconnect all the time is a waste.

Also, the tests use a :memory: connection, so in case you create a table
in one function and you try to select form it in an other function, it
won't work in the :memory: case to reconnect between the two functions.

Solving this is a bit tricky, since we pass around an immutable Context
and we want to lazy-create this (so the test Context can inject its
memory database), which means inferior mutability, twice.

Solve the lazy part by using OnceCell, we already indirectly depend on
the once_cell crate anyway. The remaining problem is how to have a
mutable connection (for transactions) inside an immutable &Context; that
can work if we hold the sql connection by Rc<RefCell<>>.

Change-Id: I1402a3692be7d148e65b151b957fe64044bd6fa0