lints, invalid-addr-cities: fix ignoring unique constraint failure
Getting the
/osm/lints/whole-country/invalid-addr-cities/update-result.json worked
the first time on a day, but later updates failed.

These non-first updates mostly worked, except that the non-update of the
stats chart resulted in an error. The non-update is intentional, but the
expected behavior is to not singal an error, either.  A problem since
commit c2cd175565c61d591fd83ee0a34aaa8ff1b46637 (stats: only ignore
SQLITE_CONSTRAINT_UNIQUE errors and still handle the rest, 2024-04-06),
which started to not ignore all error, but the constraint violated here
is a primary key one, but directly a unique one.

Fix the problem by ignoring SQLITE_CONSTRAINT_PRIMARYKEY instead of
SQLITE_CONSTRAINT_UNIQUE, which means a manual update of the stats won't
affect the recorded data for charts, and this lack of update is not
returned as an error, either.

In addition, the failure is not reported as a well-formed error, either;
that's not yet fixed here.

Change-Id: I329ffa76714ff6b0fdbe1cfe14fe5bdc8001ed2a