areas, osm housenumbers: handle the addr:unit tag
Get OSM housenumbers of a relation where the housenumber is not in a
42/b style, but addr:housenumber is 42, and addr:unit is "b". If the
reference has 42/b, then we mark 42/b as missing, which is incorrect.

The reason for this is that the overpass query gets the addr:unit tag,
but otherwise this tag is ignored.

Fix this by appending addr:unit in normalize(), so when the higher level
get_osm_housenumbers() is called for a relation, 42/b is reported to be
present in the OSM database.

Addresses <https://github.com/vmiklos/osm-gimmisn/issues/4710>.

Change-Id: I00a3abf12ef6f2e8132f7d8069393261d90a1388